Change the application type to be used in the Asgardeo console.
Issue : 
In the README.md file's getting started section, it states that creating a `Standard-Based Application` in the Asgardeo console is required. However, the next few steps are aligned with the `Traditional Web Application` rather than the `Standard-Based Application`. This is a cause of confusion. 

Fix : 
Under the  Getting started -> Create an Application in Asgardeo  section,
- Replace the `Standard-Based Application` with `Traditional Web Application` in step 2
- In step 4 remove specifying to enable Public client option.